CU-861m7wqb0 move code to directive
