+++ src/app/office/profile/edit/edit.component.scss
@@ -67,8 +67,9 @@
   }
 
   &__save {
-    position: sticky;
-    top: 0;
+    position: absolute;
+    top: 24px;
+    right: 24px;
     z-index: 10;
     display: block;
     width: 255px;
