+++ src/app/ui/directives/editor-submit-button.directive.spec.ts
@@ -1,0 +1,8 @@
+import { EditorSubmitButtonDirective } from './editor-submit-button.directive';
+
+describe('EditorSubmitButtonDirective', () => {
+  it('should create an instance', () => {
+    const directive = new EditorSubmitButtonDirective();
+    expect(directive).toBeTruthy();
+  });
+});
